refactor(utils): refactor utils into api group classes
BREAKING CHANGE: utils in ldUtils object are now grouped below type eg. flags, roles

resolves #4
